Versión 2014.4b: - Ahora se comprueba desde el principio que tengas PHP 5.3 o superior. - También se comprueba que MySQL soporte InnoDB. - Actualizado el archivo config-sample.php para poner por defecto MySQL y el remote-printer. - Añadidos más mensajes de ayuda para guiar a los novatos. - Ya se puede especificar el recargo de equivalencia para cada impuesto. - Ya se pueden imprimir albaranes de cliente en formato A4 y con logotipo. - Corregidos muchos bugs con las #etiquetas. - Ya se pueden añadir cuentas bancarias a los proveedores. - También se puede calcular el IBAN de una cuenta bancaria a partir del CCC. - Para los proveedores, además de lo ya comentado, he añadido el régimen de IVA. - En los clientes ya se puede marcar si hay que aplicarle el recargo de equivalencia. - El TPV genérico puede ahora imprimir las observaciones del ticket.
git-svn-id: http://facturascripts.googlecode.com/svn/branches/0.9@536 fe615fe8-962f-0410-811c-9ded47c8d053

+++ model/albaran_proveedor.php
@@ -427,7 +427,8 @@ class albaran_proveedor extends fs_model
       {
          $factura = new factura_proveedor();
          $factura = $factura->get($this->idfactura);
-         $factura->delete();
+         if($factura)
+            $factura->delete();
       }
       
       return $this->db->exec("DELETE FROM ".$this->table_name.
